Added WRL frontend for Win8 and WP8 Merged Qt, SDL and WRL frontend on the same corewrapper Moved OpenGL code out of SDL and Qt frontends to OpenGl backend (JGfx.cpp) Updated file system and network code to be compatible with WRL

+++ JGE/src/zipFS/zfsystem.cpp
@@ -503,7 +503,7 @@ bool filesystem::PreloadZip(const char * Filename, map<string, limited_file_info
                     continue;
 
 			    target[Name] = limited_file_info(
-				    realBeginOfFile + FileHdr.m_RelOffset,					// "Local File" header offset position
+				    (size_t)realBeginOfFile + FileHdr.m_RelOffset,					// "Local File" header offset position
 				    FileHdr.m_UncompSize					// File Size
 			    );
 		    }
